Revert "refactor: move to example"
This reverts commit 0cf6b71fbf7f7c200bdbe891a0f0462cb7ae43fa.
