+++ apps/dashboard/src/app/components/example-item/example-item.component.html
