+++ apps/dashboard/src/app/components/example-item/example-item.component.ts
@@ -2,7 +2,7 @@ import { Component, Input } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTableModule } from '@angular/material/table';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { ExampleItem } from '@app/example/models/example.item.model';
+import { ExampleItem } from '../../shared/models';
 
 @Component({
   selector: 'choh-example-item',
