+++ apps/dashboard/src/app/components/index.ts
@@ -1,6 +1,8 @@
 export * from './breadcomb/breadcomb.component';
 export * from './console-tool-bar/console-tool-bar.component';
 export * from './editor-tool-bar/editor-tool-bar.component';
+export * from './example-item/example-item.component';
+export * from './example-list/example-list.component';
 export * from './header/header.component';
 export * from './pin-assign/pin-assign.component';
 export * from './tree/tree.component';
