+++ apps/dashboard/src/app/shared/functions/index.ts
@@ -1,1 +1,2 @@
+export * from './convert';
 export * from './functions';
