+++ apps/dashboard/src/app/shared/models/example.item.model.ts
