+++ apps/dashboard/src/app/shared/models/example.model.ts
