+++ apps/dashboard/src/app/shared/service/index.ts
@@ -1,10 +1,10 @@
 // Phase 1-5 で統合されたサービス
-export * from '../../example/services/example.service';
 export * from './auth';
 export * from './chirimen';
 export * from './dialog/dialog.service';
 export * from './directory';
 export * from './editor';
+export * from './example/example.service';
 export * from './file';
 export * from './icon/icon.service';
 export * from './notification/notification.service';
